refactor: split reshard_program_factory into separate files
Split reshard_program_factory.cpp into dedicated files for better modularity:
- reshard_same_width_program_factory.cpp/.hpp - HEIGHT_SHARDED operations
- reshard_generic_program_factory.cpp/.hpp - Generic multi-core operations
- reshard_same_height_program_factory.cpp/.hpp - WIDTH_SHARDED operations

Main dispatcher function remains in reshard_program_factory.cpp. Updated
CMakeLists.txt to include new source files and fixed header declarations
for utility functions.
