+++ ttnn/cpp/ttnn/operations/data_movement/sharded/reshard/device/reshard_generic_program_factory.hpp
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ttnn/run_operation.hpp"
+
+namespace ttnn::operations::data_movement::detail {
+
+tt::tt_metal::operation::ProgramWithCallbacks reshard_multi_core_generic(const Tensor& input, Tensor& output);
+
+}  // namespace ttnn::operations::data_movement::detail
