+++ ttnn/cpp/ttnn/operations/data_movement/sharded/reshard/device/reshard_same_width_program_factory.hpp
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ttnn/run_operation.hpp"
+
+namespace ttnn::operations::data_movement::detail {
+
+template <bool is_reader>
+tt::tt_metal::operation::ProgramWithCallbacks reshard_multi_core_same_width(const Tensor& input, Tensor& output);
+
+}  // namespace ttnn::operations::data_movement::detail
